refactor: move ContextExtension tests
