+++ packages/gunshi/src/context.test.ts
@@ -1,6 +1,7 @@
 import { MessageFormat } from 'messageformat'
 import { describe, expect, test, vi } from 'vitest'
 import {
+  createMockCommandContext,
   createTranslationAdapterForIntlifyMessageFormat,
   createTranslationAdapterForMessageFormat2,
   hasPrototype
@@ -785,5 +786,45 @@ describe('CommandContextCore type', () => {
     expect(core.args).toEqual(args)
     expect(typeof core.log).toBe('function')
     expect(typeof core.translate).toBe('function')
+  })
+})
+
+describe('CommandContextExtension type', () => {
+  test('extension key is unique symbol', () => {
+    const extension1: CommandContextExtension = {
+      key: Symbol('test1'),
+      factory: () => ({ value: 1 })
+    }
+
+    const extension2: CommandContextExtension = {
+      key: Symbol('test2'),
+      factory: () => ({ value: 2 })
+    }
+
+    expect(extension1.key).not.toBe(extension2.key)
+  })
+
+  test('extension factory can return complex objects', () => {
+    const dbExtension: CommandContextExtension<{
+      // eslint-disable-next-line @typescript-eslint/no-explicit-any
+      query: (sql: string) => Promise<any>
+      transaction: (fn: () => Promise<void>) => Promise<void>
+    }> = {
+      key: Symbol('db'),
+      factory: _core => ({
+        query: async (_sql: string) => {
+          return { rows: [], count: 0 }
+        },
+        transaction: async (fn: () => Promise<void>) => {
+          await fn()
+        }
+      })
+    }
+
+    const mockCore = createMockCommandContext()
+    const db = dbExtension.factory(mockCore)
+
+    expect(typeof db.query).toBe('function')
+    expect(typeof db.transaction).toBe('function')
   })
 })
